Move CChanceState traits
