+++ include/football/CChanceState.h
@@ -1,7 +1,7 @@
 #pragma once
 
 #include "IJsonable.h"
-#include "traits/json.h"
+#include "football/traits/CChanceState.h"
 
 #include "football/types/CChanceState.h"
 #include "football/types/CGoalDrawConfiguration.h"
@@ -12,35 +12,7 @@
 
 #include "ExceptionUtils.h"
 
-namespace futsim
-{
-
-namespace football
-{
-class CChanceState;
-}
-
-template <> struct json_traits<football::CChanceState>
-{
-	//! JSON key for the class.
-	static inline constexpr std::string_view KEY = "Chance state";
-	//! JSON key for the \copybrief football::CChanceState::mChanceType
-	static inline constexpr std::string_view CHANCE_TYPE = "Chance type";
-	//! JSON key for the \copybrief football::CChanceState::mOutcome
-	static inline constexpr std::string_view OUTCOME = "Outcome";
-	/**
-	 * @brief JSON key template for the actor.
-	 * @tparam tPlayerSkill Skill used by the actor.
-	*/
-	template<football::E_PLAYER_SKILL tPlayerSkill> static inline constexpr std::string_view ACTOR{};
-};
-
-template <> inline constexpr std::string_view json_traits<football::CChanceState>::ACTOR<football::E_PLAYER_SKILL::St> = "Goalkeeper";
-template <> inline constexpr std::string_view json_traits<football::CChanceState>::ACTOR<football::E_PLAYER_SKILL::Tk> = "Tackler";
-template <> inline constexpr std::string_view json_traits<football::CChanceState>::ACTOR<football::E_PLAYER_SKILL::Ps> = "Passer";
-template <> inline constexpr std::string_view json_traits<football::CChanceState>::ACTOR<football::E_PLAYER_SKILL::Sh> = "Shooter";
-
-namespace football
+namespace futsim::football
 {
 
 /**
@@ -312,6 +284,4 @@ types::CGoalDrawConfiguration::E_CHANCE_OUTCOME DrawOutcome( const types::CChanc
 
 } //detail namespace
 
-} // football namespace
-
-} // futsim namespace
+} // futsim::football namespace
